local test of segmentation and apply mask classes
