--- flickr/config.py
@@ -1,2 +1,0 @@
-api_key = u'9430fffdb840166ec3c0a72c41dab7d7'
-api_secret = u'cf4a4010dfdcb3c8'
